Multi-language support
The main website now lives within the default /en directory for English.

Right now, because only English is supported, all users get redirected to /en regardless of their default browser language. I've included the existing .htaccess file that makes this work in htaccess-example.txt

This now allows for anyone to submit a pull request to add a translation for a new language.

I will publish a guide soon to make adding a language an easy and straightforward process, but now the required code exists to make it functional.