Fix meta text display bug
For a while, the meta description and title that a user sees when a link is shared has been displaying in the language of the lowest meta description and title on each page. This resulted in all pages' meta info displaying in a language other than English at all times, and not properly changing language based on a user's browser language.

This issue has been resolved by simply removing all meta description and meta title code that is not in English. The supermajority of web traffic is English users, so it's important those display properly.

I am not aware of a way to make translated meta descriptions and titles work with our current jquery.i18n translation setup, but I am open to options if you know of any solutions.